test(mocha): add additional tests and re-org test directory
