+++ test/fixtures/index.html
@@ -1,0 +1,2 @@
+<link href="output.css" media="all" rel="stylesheet" />
+<div class="arrow"></div>
