+++ test/fixtures/input.css
@@ -1,0 +1,33 @@
+.arrow {
+  width : 28px;
+  height : 27px;
+  margin : 10px auto;
+}
+
+.arrow {
+  background-image: url('/arrow-next.png#elements');
+}
+
+.arrow:hover {
+  background-image: url('/arrow-next_hover.png#elements');
+}
+
+@media only screen and (-webkit-min-device-pixel-ratio: 1.5) {
+  .arrow {
+    background-image: url('images/arrow-next@2x.png#elements');
+  }
+
+  .arrow:hover {
+    background-image: url('/arrow-next_hover@2x.png#elements');
+  }
+}
+
+@media only screen and (-webkit-min-device-pixel-ratio: 2.5) {
+  .arrow {
+    background-image: url('images/arrow-next@3x.png#elements');
+  }
+
+  .arrow:hover {
+    background-image: url('images/arrow-next_hover@3x.png#elements');
+  }
+}
