+++ test/fixtures/output.css
@@ -1,0 +1,27 @@
+.arrow {
+  width : 28px;
+  height : 27px;
+  margin : 10px auto;
+}
+
+.arrow { background-image : url(sprites/elements.png); background-position : 0 0;
+}
+
+.arrow:hover { background-image : url(sprites/elements.png); background-position : -48px 0;
+}
+
+@media only screen and (-webkit-min-device-pixel-ratio: 1.5) {
+  .arrow { background-image : url(sprites/elements@2x.png); background-position : 0 0; background-size : 66px 27px;
+  }
+
+  .arrow:hover { background-image : url(sprites/elements@2x.png); background-position : -38px 0; background-size : 66px 27px;
+  }
+}
+
+@media only screen and (-webkit-min-device-pixel-ratio: 2.5) {
+  .arrow { background-image : url(sprites/elements@3x.png); background-position : 0 0; background-size : 62.666666666666664px 27px;
+  }
+
+  .arrow:hover { background-image : url(sprites/elements@3x.png); background-position : -34.666666666666664px 0; background-size : 62.666666666666664px 27px;
+  }
+}
