+++ test/fixtures/reference-images/spriteselements@3x.png
